dev_tree: Add local-bd-address property to BT node
A local-bd-address is now expected by the Bluetooth driver
in order to set the bluetooth controller unique address.
Contrary to local-mac-address, the local-bd-address must
be formatted with least significant byte first (little-endian).

Keep the local-mac-address BT node property in order to guaranty
compatibility with existing user-space scripts and non-patched
kernel.

Change-Id: I91779009194977f619f0ad6c45117b8d3cd8288b
Signed-off-by: Loic Poulain <loic.poulain@linaro.org>
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>